Fix input on padded dropdown search bars
